requirements.txt
pytorch-lightning>=1.0.8,<2
scikit-multilearn>=0.2.0,<1
scipy>=1.5.4,<2
pyyaml>=5.3.1,<6
pandas>=1.1.4,<2
dask>=2.30.0,<3
scikit-learn>=0.23.2,<1
imbalanced-learn>=0.7.0,<1
numpy>=1.19.4,<2
requests>=2.22.0,<3
pydantic>=1.6.1,<2
